requirements.txt
notion-client
line-bot-sdk
fastapi
deta